pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "A framework for building robust, extensible, and maintainable data processing pipelines."
dependencies = [
    "pandas",
    "numpy",
    "PyYAML",
    "Pillow",
    "typer[all]",
    "pytest",
    "pytest-dependency",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
data-replay = "nexus.cli:app"

